+++ src/types/index.ts
@@ -1,0 +1,22 @@
+export interface Structure {
+    name: string;
+    variables?: Variable[];
+    optionals?: Optional[];
+    structure: StructureItem[];
+}
+
+export interface Variable {
+    varName: string;
+    default?: string;
+}
+
+export interface Optional {
+    optName: string;
+    value?: boolean;
+}
+
+export interface StructureItem {
+    fileName: string;
+    template?: string;
+    optional?: string;
+}