+++ src/utils/pathUtils.ts
@@ -20,7 +20,8 @@ export async function getTargetPath(Uri?: vscode.Uri): Promise<{ targetPath: str
         if (parentFolderUri?.[0]) {
             targetPath = parentFolderUri[0].fsPath;
             createNewFolder = true;
-        } else {
+        }
+        else {
             vscode.window.showErrorMessage('No folder selected', { modal: true });
         }
     }